+++ brunch-config/lib.js
@@ -1,72 +1,48 @@
+require('sugar');
 var basename = require('path').basename;
 
-// Add a number of tests to the ignored. Tests can be any of the following:
-//   Function: Return true to ignore
-//   RegExp:   Match to ignore
-//   String:   Equal to ignore
+/**
+ * Add additional constraints to Brunch's ignored setting using tests. Tests
+ * can be any of the following:
+ *    Function: Return true to ignore
+ *    RegExp:   Match to ignore
+ *    String:   Equal to ignore
+ * @param {Object} config
+ * @param {...(Function|RegExp|String)} tests
+ */
 exports.addIgnored = function(config) {
   if(!config.conventions) {
     config.conventions = {};
   }
+
+  // Get original ignored function and all given tests
   var ignored = config.conventions.ignored;
   var tests = Array.prototype.slice.call(arguments, 1);
 
+  // Set up the new ignored function
   config.conventions.ignored = function(file) {
+    // Check if any of the given tests pass
     for(var index = 0, test = tests[index]; index < tests.length; index++) {
       if(test === file) {
         return true;
       }
-      if((typeOf(test) === 'Function') && test(file)) {
+      if(Object.isFunction(ignored) && test(file)) {
         return true;
       }
-      if((typeOf(test) === 'RegExp') && test.test(file)) {
+      if(Object.isRegExp(ignored) && test.test(file)) {
         return true;
       }
     }
 
-    if(typeOf(ignored) === 'Function') {
+    // Otherwise, delegate to the original ignored declaration
+    if(Object.isFunction(ignored)) {
       return ignored(file);
     }
-    if(typeOf(ignored) === 'RegExp') {
+    if(Object.isRegExp(ignored)) {
       return ignored.test(file);
     }
+
+    // This is the default test if any of the previous tests fail
     return basename(file).indexOf('_') === 0;
   };
 }
-
-// Replacement for typeof keyword for slightly more detailed information
-var typeOf = exports.typeOf = function(obj) {
-  return Object.prototype.toString.call(obj).slice(8, -1);
-};
-
-// Deep clone an object (usually for the Brunch configuration object)
-var clone = exports.clone = function(original) {
-  switch(typeOf(original)) {
-    case 'Object':
-      var object = {};
-      for(var key in original) {
-        if(original.hasOwnProperty(key)) {
-          object[key] = clone(original[key]);
-        }
-      }
-      return object;
-
-    case 'Array':
-      return original.map(function(value) {
-        return clone(value);
-      });
-
-    case 'RegExp':
-      return new RegExp(original);
-
-    case 'Date':
-      return new Date(original.getTime());
-
-    case 'Function':
-      return function() {
-        return original.apply(null, arguments);
-      };
-  }
-
-  return original;
-};