+++ brunch-config/platform.js
@@ -1,4 +1,9 @@
-// Adjust Brunch configuration object to fit a specific platform
+/**
+ * Modify given Brunch configuration to fit a given platform.  Currently
+ * there is support for a web platform.
+ * @param  {String} platform
+ * @param  {Object} config
+ */
 module.exports = function(platform, config) {
   switch(platform) {
     case 'web':
@@ -7,13 +12,20 @@ module.exports = function(platform, config) {
   }
 };
 
-// List of platforms supported
+/**
+ * List of platforms supported.
+ * @type {Array}
+ */
 module.exports.platforms = ['web'];
 
-// For a generic web app, simply specify public directory
+/**
+ * Modify configuration to match the web platform. This entails setting the
+ * public directory path.
+ * @param  {Object} config
+ */
 function webPlatform(config) {
   if(!config.paths) {
     config.paths = {};
   }
-  config.paths.public = 'build/web';
+  config.paths.public = 'public';
 }