+++ brunch-config/setup.js
@@ -1,22 +1,28 @@
-var clone = require('./lib').clone;
+require('sugar');
 var setPlatform = require('./platform');
 var setEnvironment = require('./environment');
 
-// Set up config by taking the given base configuration and split
-// it into multiple pieces based on overrides.
-module.exports = function(config) {
-  var newConfig = {
+/**
+ * Take a given Brunch configuration and append additional configuration
+ * options for different environments and platforms. For more information on
+ * that see environment.js and platform.js.
+ * @param  {Object} originalConfig
+ * @return {Object}
+ */
+module.exports = function(originalConfig) {
+  var finalConfig = {
     overrides: {}
   };
 
+  // Generate full config clones for each combination of platform/environment.
   setPlatform.platforms.forEach(function(platform) {
     setEnvironment.environments.forEach(function(environment) {
-      var conf = clone(config);
-      setPlatform(platform, conf);
-      setEnvironment(environment, conf);
-      newConfig.overrides[platform + ':' + environment] = conf;
+      var config = Object.clone(originalConfig, true);
+      setPlatform(platform, config);
+      setEnvironment(environment, config);
+      finalConfig.overrides[platform + ':' + environment] = config;
     });
   });
 
-  return newConfig;
+  return finalConfig;
 }