+++ test/code/sample-test.js
@@ -1,4 +1,4 @@
-describe('Sample test', function() {
+describe('Sample', function() {
   it('2 + 2 = 4', function() {
     expect(2 + 2).to.equal(4);
   });