+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 - Convert custom Cake tasks into standard Jake tasks
 - Remove any prompts in favor of Jake arguments
 - Remove `npm run-script` tasks. Use `jake`.
-- Add Mocha PhantomJS support for whitebox testing
-- Add Lotte support for blackbox testing
-- Add `auto-reload brunch` by default
+- Auto generate Jake tasks for Scaffolt generators
+- Add code testing and site testing support
+- Add PhantomJS for testing
+- Add `auto-reload-brunch` by default
 - Add test file generator
 - Remove Modernizr
+- Use default `public` path for default builds
 
 #### 0.3.2 (September 27, 2013)
 - Add ability to edit `bower.json` in bower task